Moving flyPad translations to the new fbw-a32nx/localisation folder
