+++ fbw-a32nx/src/localisation/.eslintrc
