+++ fbw-a32nx/src/localisation/flypad/README.md
@@ -1,0 +1,14 @@
+# flyPad Localization
+
+Only change the `en.json` file. 
+
+The other files are generated automatically from Localazy and any changes will be overwritten.
+                                
+[//]: # (TODO: GitHub secrets do not work on PRs from forks - check if this workflow is even possible)
+
+Add new keys to the `en.json` file. These will be automatically added to Localazy and will be available for translation
+after the PR has been approved. 
+
+Changes to existing keys will be automatically updated in Localazy once the PR is merged to master. They will be visible 
+in the PR build though.  
+
