+++ fbw-a32nx/src/localisation/flypad/es.json
