+++ fbw-a32nx/src/localisation/flypad/fr.json
