+++ fbw-a32nx/src/localisation/flypad/hi.json
