+++ fbw-a32nx/src/localisation/flypad/ja.json
