+++ fbw-a32nx/src/localisation/flypad/ko.json
