+++ fbw-a32nx/src/localisation/flypad/localazy-flypad-upload-deprecate-config.json
@@ -1,0 +1,17 @@
+{
+  "upload": {
+    "skipFailedAutodetection": true,
+    "deprecate": "project",
+    "appVersion": 0,
+    "keySeparator": ".",
+    "features": [],
+    "folder": ".",
+    "type": "json",
+    "files": {
+      "pattern": "en.json",
+      "type": "json",
+      "path": ".",
+      "file": "flybywireos3.json"
+    }
+  }
+}
