+++ fbw-a32nx/src/localisation/flypad/pl.json
