+++ fbw-a32nx/src/localisation/flypad/pt-BR.json
