+++ fbw-a32nx/src/localisation/flypad/pt-PT.json
