+++ fbw-a32nx/src/localisation/flypad/ru.json
