+++ fbw-a32nx/src/localisation/flypad/sk.json
