+++ fbw-a32nx/src/localisation/flypad/zh-Hans-CN.json
