--- fbw-a32nx/src/systems/instruments/src/EFB/Localization/localazy.json
@@ -1,22 +1,0 @@
-{
-  "upload": {
-    "skipFailedAutodetection": true,
-    "importAsNew": false,
-    "forceCurrent": false,
-    "appVersion": 0,
-    "keySeparator": ".",
-    "features": [],
-    "folder": ".",
-    "type": "json",
-    "files": [
-      "flybywireos3.json"
-    ]
-  },
-  "download": {
-    "folder": "downloaded",
-    "includeSourceLang": true,
-    "files": {
-      "output": "${lang}.json"
-    }
-  }
-}
