+++ fbw-a32nx/src/systems/instruments/src/EFB/translation.ts
@@ -6,42 +6,42 @@
 import { NXDataStore } from '@shared/persistence';
 
 // source language
-import en from './Localization/en.json';
+import en from '../../../../localisation/flypad/en.json';
 // translations
-import ar from './Localization/ar.json';
-import cs from './Localization/cs.json';
-import de from './Localization/de.json';
-import da from './Localization/da-DK.json';
-import el from './Localization/el.json';
-import es from './Localization/es.json';
-import eu from './Localization/eu.json';
-import fi from './Localization/fi.json';
-import fr from './Localization/fr.json';
-import he from './Localization/he.json';
-import hi from './Localization/hi.json';
-import hu from './Localization/hu.json';
-import hr from './Localization/hr.json';
-import id from './Localization/id.json';
-import it from './Localization/it.json';
-import ja from './Localization/ja.json';
-import ko from './Localization/ko.json';
-import lt from './Localization/lt.json';
-import nb from './Localization/nb.json';
-import nl from './Localization/nl.json';
-import pl from './Localization/pl.json';
-import ptBR from './Localization/pt-BR.json';
-import ptPT from './Localization/pt-PT.json';
-import ro from './Localization/ro.json';
-import ru from './Localization/ru.json';
-import sk from './Localization/sk.json';
-import sl from './Localization/sl.json';
-import sv from './Localization/sv.json';
-import th from './Localization/th.json';
-import tr from './Localization/tr.json';
-import vi from './Localization/vi.json';
-import zhHansCN from './Localization/zh-Hans-CN.json';
-import zhHantHK from './Localization/zh-Hant-HK.json';
-import zhHantTW from './Localization/zh-Hant-TW.json';
+import ar from '../../../../localisation/flypad/ar.json';
+import cs from '../../../../localisation/flypad/cs.json';
+import de from '../../../../localisation/flypad/de.json';
+import da from '../../../../localisation/flypad/da-DK.json';
+import el from '../../../../localisation/flypad/el.json';
+import es from '../../../../localisation/flypad/es.json';
+import eu from '../../../../localisation/flypad/eu.json';
+import fi from '../../../../localisation/flypad/fi.json';
+import fr from '../../../../localisation/flypad/fr.json';
+import he from '../../../../localisation/flypad/he.json';
+import hi from '../../../../localisation/flypad/hi.json';
+import hu from '../../../../localisation/flypad/hu.json';
+import hr from '../../../../localisation/flypad/hr.json';
+import id from '../../../../localisation/flypad/id.json';
+import it from '../../../../localisation/flypad/it.json';
+import ja from '../../../../localisation/flypad/ja.json';
+import ko from '../../../../localisation/flypad/ko.json';
+import lt from '../../../../localisation/flypad/lt.json';
+import nb from '../../../../localisation/flypad/nb.json';
+import nl from '../../../../localisation/flypad/nl.json';
+import pl from '../../../../localisation/flypad/pl.json';
+import ptBR from '../../../../localisation/flypad/pt-BR.json';
+import ptPT from '../../../../localisation/flypad/pt-PT.json';
+import ro from '../../../../localisation/flypad/ro.json';
+import ru from '../../../../localisation/flypad/ru.json';
+import sk from '../../../../localisation/flypad/sk.json';
+import sl from '../../../../localisation/flypad/sl.json';
+import sv from '../../../../localisation/flypad/sv.json';
+import th from '../../../../localisation/flypad/th.json';
+import tr from '../../../../localisation/flypad/tr.json';
+import vi from '../../../../localisation/flypad/vi.json';
+import zhHansCN from '../../../../localisation/flypad/zh-Hans-CN.json';
+import zhHantHK from '../../../../localisation/flypad/zh-Hant-HK.json';
+import zhHantTW from '../../../../localisation/flypad/zh-Hant-TW.json';
 
 console.log('Initializing Translation');
