+++ package.json
@@ -8,7 +8,7 @@
     "build-a32nx:copy-base-package": "mkdir -p fbw-a32nx/out/flybywire-aircraft-a320-neo && (rsync -a fbw-a32nx/src/base/flybywire-aircraft-a320-neo fbw-a32nx/out/ || cp -a -u fbw-a32nx/src/base/flybywire-aircraft-a320-neo fbw-a32nx/out/)",
     "build-a32nx:copy-base-lock-highlight-package": "mkdir -p fbw-a32nx/out/flybywire-aircraft-a320-neo-lock-highlight && (rsync -a fbw-a32nx/src/base/flybywire-aircraft-a320-neo-lock-highlight fbw-a32nx/out/ || cp -a -u fbw-a32nx/src/base/flybywire-aircraft-a320-neo-lock-highlight fbw-a32nx/out/)",
     "build-a32nx:copy-base-files": "npm run build-a32nx:copy-base-package && npm run build-a32nx:copy-base-lock-highlight-package",
-    "build-a32nx:efb-translation": "cd fbw-a32nx/src/systems/instruments/src/EFB/Localization && node build-translation.js",
+    "build-a32nx:efb-translation": "cd fbw-a32nx/src/localisation && node build-flypad-translation.js",
 
     "build-a32nx:model": "node fbw-a32nx/src/model/build.js",
     "build-a32nx:behavior": "node fbw-a32nx/src/behavior/build.js",
